说明: + 增加工作流程 of::work 独立的 时间 队列 错误及事务, 让代码更简洁       + 框架开始兼容 php 7.4       + 框架控制台增加运行环境信息       + of_db::pool 增加"rename"连接池更名, 更名会触发"of_db::rename"事件       + of_db::pool 增加"clone"连接池克隆, 克隆名称冲突时, 原连接会触发更名成唯一值       + of_db::pool 增加"ping"检查连接池是否连接, 已连接返回true, 未连接返回false       + 错误格式标准化 {"code" : 编码, "info" : 错误, "file" : 路径, "line" : 行数, ...}       - 数据库错误接口 of_accy_db_xxx::_error 返回结构变更       - 数据库添加接口 of_accy_db_xxx::_ping 检查连接是否正常       - 事件 of_base_error_writeLog::error 触发参数符合错误格式       - 事件 of_db::error 触发参数 type 分解为 code 和 info       - of::syntax 返回参数 message 变更为 info       * 用 of::work('error') 更好的代替 of_base_error_writeLog::lastError()       * 解决 of_base_com_com::arraySort 对 $data[""] 排序后变为 $data[0] 的问题       * 解决事件 of::loadClass 通过 mapping 加载命名空间类寻址错误问题       * 更有序的嵌套事件监听触发顺序, 同时解决由相同嵌套事件回调导致的死循环       * of_db::pool 销毁连接时, 触发相同事务层数的回滚操作       * 方法 L::getText 改为 L 默认方法

+++ demo/model/list.php
@@ -7,21 +7,21 @@ class ctrl_list {
     public static function init() {
         //ctrl_类映射
         of::event('of::loadClass', array(
-            'classPre' => 'ctrl_', 'mapping' => 'demo_ctrl_'
+            'classPre' => 'ctrl', 'mapping' => 'demo_ctrl'
         ), true);
         //model_类映射
         of::event('of::loadClass', array(
-            'classPre' => 'model_', 'mapping' => 'demo_model_'
+            'classPre' => 'model', 'mapping' => 'demo_model'
         ), true);
         //serv_类映射
         of::event('of::loadClass', array(
-            'classPre' => 'serv_', 'mapping' => 'demo_serv_'
+            'classPre' => 'serv', 'mapping' => 'demo_serv'
         ), true);
 
         //非框架入口
         if ($_SERVER['SCRIPT_NAME'] === rawurldecode(ROOT_URL) . '/index.php') {
             //具体演示
-            if( isset($_GET['c']) ) {
+            if (isset($_GET['c'])) {
                 //页面输出完成后打印源码
                 empty($_POST) && strncmp('of_', $_GET['c'], 3) && of::event('of::halt', 'printCode');
             //打印演示列表
@@ -32,7 +32,7 @@ class ctrl_list {
 
                 //打印html头
                 of_view::head();
-                foreach($match as &$v) {
+                foreach ($match as &$v) {
                     //打印演示列表
                     echo "<a href='?c=ctrl_index&a={$v[2]}' target=_blank >{$v[1]}</a><br>\n";
                 }
@@ -47,9 +47,14 @@ class ctrl_list {
             function printCode() {
                 //提取方法体
                 $matchFunc = '@^( +)public +(?:\w+ +)?function +' .(isset($_GET['a']) ? $_GET['a'] : 'index'). '.*?^\\1}@ms';
-                preg_match($matchFunc, file_get_contents(ROOT_DIR . '/demo/ctrl/index.php'), $match);
-                echo '<br><br><hr>';
-                highlight_string("<?php    源码如下 : \n" . $match[0] . "\n?>");
+                $path = ROOT_DIR . '/demo/' . str_replace(array('\\', '_'), '/', of::dispatch('class')) . '.php';
+
+                //解析文件存在
+                if (is_file($path)) {
+                    preg_match($matchFunc, file_get_contents($path), $match);
+                    echo '<br><br><hr>';
+                    highlight_string("<?php    源码如下 : \n" . $match[0] . "\n?>");
+                }
             }
         }
     }
